YAY! Got the AI working!
There does seem to be a problem of the AI getting things in the wrong priority
order. Going 4 to 1 gives the wrong move and 1 to 4 sometimes gives the wrong
answer. But it does make moves. Theres still no win condition however and
only the first row of black chips is spawned. But the AI is done!